Normal problems resolution: 1. Adding control points causes weird shapes to form  - ask Damon to remove "if (!this->path->vertex_data)" and still restrict memory usage 2. Suddenly upside-down control point causes track to break  - use slight slopes in game

+++ GraphicsRush/src/GMTPipeline.cpp
@@ -18,15 +18,15 @@ void GMTPipeline::setG_pos(int cp_id) {
 };
 void GMTPipeline::setG_orient(int cp_id) {
 	G = {
-			{gmt_track->points[(cp_id - 1 + gmt_track->points.size()) % (int)gmt_track->points.size()].orient.x,
+			{gmt_track->points[(cp_id - 1 + (int)gmt_track->points.size()) % (int)gmt_track->points.size()].orient.x,
 			gmt_track->points[cp_id % (int)gmt_track->points.size()].orient.x,
 			gmt_track->points[(cp_id + 1) % (int)gmt_track->points.size()].orient.x,
 			gmt_track->points[(cp_id + 2) % (int)gmt_track->points.size()].orient.x},
-			{ gmt_track->points[(cp_id - 1 + gmt_track->points.size()) % (int)gmt_track->points.size()].orient.y,
+			{ gmt_track->points[(cp_id - 1 + (int)gmt_track->points.size()) % (int)gmt_track->points.size()].orient.y,
 			gmt_track->points[cp_id % (int)gmt_track->points.size()].orient.y,
 			gmt_track->points[(cp_id + 1) % (int)gmt_track->points.size()].orient.y,
 			gmt_track->points[(cp_id + 2) % (int)gmt_track->points.size()].orient.y },
-			{ gmt_track->points[(cp_id - 1 + gmt_track->points.size()) % (int)gmt_track->points.size()].orient.z,
+			{ gmt_track->points[(cp_id - 1 + (int)gmt_track->points.size()) % (int)gmt_track->points.size()].orient.z,
 			gmt_track->points[cp_id % (int)gmt_track->points.size()].orient.z,
 			gmt_track->points[(cp_id + 1) % (int)gmt_track->points.size()].orient.z,
 			gmt_track->points[(cp_id + 2) % (int)gmt_track->points.size()].orient.z },
